+++ include/H2DE/engine/H2DE_lerp_manager.h
@@ -7,7 +7,7 @@ class H2DE_Engine;
 class H2DE_LerpManager {
 public:
     template<typename H2DE_Lerp_T>
-    static unsigned int lerp(H2DE_Engine* engine, H2DE_Lerp_T origin, H2DE_Lerp_T destination, unsigned int duration, H2DE_Easing easing, const std::function<void(H2DE_Lerp_T)>& update, const std::function<void()>& completed, bool pauseSensitive) {
+    static unsigned int lerp(H2DE_Engine* engine, H2DE_Lerp_T origin, H2DE_Lerp_T destination, unsigned int duration, H2DE_Easing easing, const std::function<void(H2DE_Lerp_T)>& update, const std::function<void()>& completed, bool pauseSensitive = true) {
         const H2DE_Lerp_T valueToAdd = destination - origin;
 
         return engine->createTimeline(duration, easing, [update, origin, valueToAdd](float blend) {
@@ -18,7 +18,7 @@ public:
         }, completed, 0, pauseSensitive);
     }
 
-    static unsigned int lerp(H2DE_Engine* engine, const H2DE_ColorRGB& origin, const H2DE_ColorRGB& destination, unsigned int duration, H2DE_Easing easing, const std::function<void(H2DE_ColorRGB)>& update, const std::function<void()>& completed, bool pauseSensitive) {
+    static unsigned int lerp(H2DE_Engine* engine, const H2DE_ColorRGB& origin, const H2DE_ColorRGB& destination, unsigned int duration, H2DE_Easing easing, const std::function<void(H2DE_ColorRGB)>& update, const std::function<void()>& completed, bool pauseSensitive = true) {
         const Uint8 redToAdd = destination.r - origin.r;
         const Uint8 greenToAdd = destination.g - origin.g;
         const Uint8 blueToAdd = destination.b - origin.b;