+++ include/H2DE/surfaces/H2DE_color.h
@@ -20,7 +20,7 @@ public:
 
     inline void setColor(const H2DE_ColorRGB& color) { colorData.color = color; }
 
-    unsigned int setColor(const H2DE_ColorRGB& color, unsigned int duration, H2DE_Easing easing, const std::function<void()>& completed, bool pauseSensitive);
+    unsigned int setColor(const H2DE_ColorRGB& color, unsigned int duration, H2DE_Easing easing, const std::function<void()>& completed, bool pauseSensitive = true);
 
     friend class H2DE_Object;
 };