+++ include/H2DE/surfaces/H2DE_texture.h
@@ -22,7 +22,7 @@ public:
     void setColor(const H2DE_ColorRGB& color);
     void setSrcRect(const std::optional<H2DE_PixelRect>& srcRect);
 
-    unsigned int setColor(const H2DE_ColorRGB& color, unsigned int duration, H2DE_Easing easing, const std::function<void()>& completed, bool pauseSensitive);
+    unsigned int setColor(const H2DE_ColorRGB& color, unsigned int duration, H2DE_Easing easing, const std::function<void()>& completed, bool pauseSensitive = true);
 
     friend class H2DE_Object;
     friend class H2DE_TextObject;